requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.0.0
anthropic>=0.10.0
gradio>=4.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
